Changes for navbar and options panel
